minor changes
changed so all responses back to Slack is consistent (previously it was inconsistent - JSON response, fmt.Fprint and http.ResponseWriter.Write)
added an option to specify sub-teamID of Slack's default admin group "@admins" - this is used in error message when external services (Slack, Google Datastore) had an problem and the text says "please contact @admins". without the sub-teamID, it is a pure text. With sub-teamID this will become a Slack mention (and clickable)